src/backend/requirements.txt
fastapi[all]==0.70.1
uvicorn[standard]==0.15.0
gunicorn==20.1.0
httpx==0.23.3 
python-dotenv==0.21.1
psycopg2==2.9.3
sqlalchemy==2.0.4